--- ep/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "ep",
-  "license": "UNLICENSED",
-  "scripts": {
-    "codegen": "graph codegen",
-    "build": "graph build",
-    "deploy": "graph deploy --node https://api.studio.thegraph.com/deploy/ ep",
-    "create-local": "graph create --node http://localhost:8020/ ep",
-    "remove-local": "graph remove --node http://localhost:8020/ ep",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 ep",
-    "test": "graph test"
-  },
-  "dependencies": {
-    "@graphprotocol/graph-cli": "0.95.0",
-    "@graphprotocol/graph-ts": "0.36.0"
-  },
-  "devDependencies": { "matchstick-as": "0.6.0" }
-}
