--- ep/schema.graphql
@@ -1,57 +1,0 @@
-type LogicContractSet @entity(immutable: true) {
-  id: Bytes!
-  _version: String! # string
-  _upgrade: BigInt! # uint256
-  _logicContract: Bytes! # address
-  _initializationData: Bytes! # bytes
-  _upgradeData: Bytes! # bytes
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
-
-type TokenUpgraded @entity(immutable: true) {
-  id: Bytes!
-  _securityToken: Bytes! # address
-  _version: BigInt! # uint256
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
-
-type DefaultTransferManagerUpdated @entity(immutable: true) {
-  id: Bytes!
-  _oldTransferManagerFactory: Bytes! # address
-  _newTransferManagerFactory: Bytes! # address
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
-
-type DefaultDataStoreUpdated @entity(immutable: true) {
-  id: Bytes!
-  _oldDataStoreFactory: Bytes! # address
-  _newDataStoreFactory: Bytes! # address
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
-
-type OwnershipTransferred @entity(immutable: true) {
-  id: Bytes!
-  previousOwner: Bytes! # address
-  newOwner: Bytes! # address
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
-
-type LogicContractSet1 @entity(immutable: true) {
-  id: Bytes!
-  _version: String! # string
-  _logicContract: Bytes! # address
-  _upgradeData: Bytes! # bytes
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
